[FIX] l10n_th_reports: file export with the wizard
The file export for the two pnd csv files works as
expected when using their button, but not when
using the save wizard.

This is because the data is not returned as bytes
while it is what is expected by the wizard.
Changing it still works while exporting directly
but also fixes the issue in the wizard.

Also change the file_export_type of the tax report
exports so that they have can be more easily identified
in the save wizard.

closes odoo/enterprise#45701

X-original-commit: 27113fb7c04cc9cb7099fb1d4593865b9bd959b3
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>