[REF] spreadsheet_edition: easier copy access right
`spreadsheet.revision` model is a technical model for spreadsheet
collaboration. It represents every operation done on a spreadsheet.
(If you have the revisions, you can rebuild the entire spreadsheet)

Concerning the access rights, it must be the exact same as the spreadsheet.
If the user can't read the spreadsheet, he can't read the revisions.
If the user can't write the spreadsheet, he can't write the revisions.
etc.

This kind of access right delegation is not possible with the current
framework. It must be handled by hand.

The chosen strategy was to restrict all access to the revision model
and grant access with `sudo` on a case by case basis, after explicit
checking.
It's easier/safer to keep the door closed by default and open it
when needed, rather than making sure we didn't forget to protect a
window.

However, that means a user can read `document.spreadsheet_revision_ids`
even if he has read access on the document. `sudo` must be used.

Currently, `sudo` was required to copy a spreadsheet's revisions.
So everytime `_copy_revisions_to` was used, it was done with `sudo`.
It's ok if the caller makes sure it's safe.
But if the method is used more and more, it means more and more
`sudo` calls, and more and more risk of forgetting to check.

This commit moves the `sudo` calls to the method itself, so the caller
don't have to worry about it.
An explicit check is done in the method itself.
Current checks were implicit btw, which is not a good idea. While it is
safe at the moment, code evolution could break it if we are not
very careful.

closes odoo/enterprise#45676

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>