[FIX] l10n_nl_reports: file export with the wizard
The file export for the XAF files works as
expected when using their button, but not when
using the save wizard.

This is because the data is not returned as bytes
while it is what is expected by the wizard.
Changing it still works while exporting directly
but also fixes the issue in the wizard.

Also change the file_type of the tax report
exports so that they keep their extension when saved in Documents via
the wizard.

closes odoo/enterprise#45952

X-original-commit: f31f3e4432ce1a1b5e9f6b477af00155092cddaf
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>