[FIX] hr_holidays_gantt : fix _group_leaves()
To reproduce the bug:

- First, consider a week spanning from the 18th (Monday) to the 22nd
(Friday).
- Next, input validated time off for the dates 18th through the 20th.
- Then, input validated time off specifically for the 22nd.
- After that, make a request for time off on the 21st.
- Finally, assign a task to the user for the entire week: 18th to 22nd.

The system's warning only highlights the time off from the 18th to the
20th and the 22nd. It misses the requested time off on the 21st.
The root of this problem lies in the _group_leaves() function. If a time
off request is chronologically after a validated time off in the list,
but is earlier in terms of 'time' (like the example provided), it isn't
considered.

To address this, when merging time off entries, ensure to verify if the
current time off type is different from the preceding one.

task-3330290

Part-of: odoo/enterprise#42537