[FIX] industry_fsm: avoid translating template logic in German
The German translation file translates logic in master which leads to
traceback when rendering those templates.

Steps
=====
- Install German language (but keep the db in english)
- Create a contact with German as a language
- In FSM, go to configuration > stages
- Open the stage "New" form view
- In the email template field, choose: "Task: Intervention Scheduled"
- Create a task with the created contact as customer

Issue
=====
A traceback appears indicating that the template could not be rendered.

Cause
=====
Some template logic has been translated in German in the .po file.

Fix
===
In stable, the modification have been done in the odoo projects on
transifex. In master the project is not yet existing. Therefore the
modifications have directly been done in the relevant files.

task-3447430

closes odoo/enterprise#45688

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>