[FIX] account_followup: test fails undeterministically
The test test_onchange_residual_amount fails undeterministically.
The reason being that the unreconciled_aml_ids have no order defined
in the Form. It is defined in the xml view, but the mocking does not
take it into account.
We just define the line to block with the date.

Linked to runbot error 24031

closes odoo/enterprise#45722

Signed-off-by: Laurent Smet (las) <las@odoo.com>