[IMP] account_accountant: sort bank rec AML's by matching amount
When reconciling bank transactions, it is ideal to have a suggested AML found by the reco model.
But in cases where there is no partner set on the transaction, a suggestion from the bank reco model is not possible.
To simplify the recon process, we display AML's that have a matching amout_residual first.

In order to achieve this a custom ORDER BY is used.
Since this ORDER BY results in a performance hit, it is adapted to make use of an index, thereby minimising the performance impact.
The index creation is therefore moved from account_followup to account.

Task-3042547

closes odoo/enterprise#45816

Related: odoo/odoo#135155
Related: odoo/upgrade#5146
Signed-off-by: Laurent Smet (las) <las@odoo.com>