[FIX] documents_project_sale : create product with company's workspace
Issue:
======
When you select a workspace for a product that has been created by a
project it will throws an error showing that it has a refrence for a
False Company.

Steps to reproduce the error:
=============================
1-Install sales, documents, project
2-Create a project (named test)
3-Go to sales/products create new product
4-Select type as service, create on order : project & task, project
template : test, workspace template : projects/test
5- click save and it will show the error

Solution:
=========
We should check that product has a company_id.

opw-3456166

closes odoo/enterprise#45927

X-original-commit: 37c3ff6009055ae00756385c4848d76f778e898a
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>