[IMP] knowledge: adapt dragAndDrop util to `useNestedSortable` changes
Purpose:
--------
The `useNestedSortable` hook has been improved in the related community
PR. One of these improvements consists in using the position of the top
of the dragged element instead of the position of the cursor to detect
if a move should be triggered.
The `dragAndDropArticle` util used in the test tours has been updated to
reflect this change.

Task-3422012

Part-of: odoo/enterprise#45729