[FIX] mrp_workorder: assign serial not working on workorder
On the shop floor app. In a workorder context. On the step register
production. The assign serial button was correctly doing the rpc but
doesn't update the record in the client.

It happens because the finished serial is on the production. And the
mrp display record only reload itself and its subcomponents. However
the production is upper and should be reload in that case

X-original-commit: c88635495e9c3efb64b654e8b0c59165cde0e2ab
Part-of: odoo/enterprise#45932