[FIX] helpdesk: get the right sequence on ticket in multi-company environment
Current behaviour:
In helpdesk in a multiple company configuration, when we send a mail to a team
in a company it always creates a ticket with the sequence of the default
company.

Steps to reproduce:
- Enable 'Custom email servers' in settings, set the domain and add a company.
- Create 2 helpdesk teams in each company with a specific alias for the emails.
- Go to Sequences, set two of them with 'helpdesk.ticket' as sequence code and different
prefixes and companies to match your teams.
- Email with the alias created for the team in the added company(you can use this lines of codes
to do it locally https://github.com/odoo/odoo/blob/bfdd54e815334dbfd0f65ec6559a711b4973a6eb/addons/crm/tests/test_crm_lead_notification.py#L102-L123,
then run ``self.env.cr.commit()``)
- Go to Helpdesk in the added commpany and observe the created ticket

Issue:
The sequence used is the one of the first company.

Reason of the problem:
When creating the ticket we use the ``self.env.company`` to match the sequence,
so when it's from a email it's always the default company which is used. We have to use
the company of the ``Helpdesk.team`` to create it.

opw-3345347

closes odoo/enterprise#45971

X-original-commit: f3fe033828a7de07c3343b1ea228054c09d7c546
Signed-off-by: Eteil Junior Djoumatchoua (etdj) <etdj@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>