[FIX] mrp_workorder: shop floor don't start workorders from MO cards
Before this commit, clicking on a work order line on an MO card in the
new shop floor display would start/stop work on the workorder, which was
confusing to some users.
This commit fixes this by removing this behavior (long-press and the
workcenter link still work).

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209