[FIX] mrp_workorder: visual and functional changes
== Keep the move button clickable ==

Before this commit, the button to mark the move's quantity done as
fulfilled was disabled once the demand is completed.
This commit removes the "disabled" CSS class and replaces it by other
CSS classes to keep the same style while still be able to click again on
this button.
Also for the `StockMove` component, renames `openMoveDetails` into
`toggleQuantityDone` since this method doesn't open the detailed
operation view (but the method `clicked` does !) and set/unset the
quantity done instead.

== Change the way Unit/Units is displayed when UoM setting is off ==

Before this commit, when UoM setting was off, "Unit" was always
displayed and a "s" was added at the end if multiple quantities.
But doing this is not recommanded for translation. So, the complete word
for "Unit" or "Units" is given to the translation method.

== Quality Checks' labels ==

- Display nothing if QC has no title (previously display "False");
- Use the right label depending of the QC product's tracking (always
  displayed the finished product's tracking before this commit).

task-3456817

Part-of: odoo/enterprise#46701