[FIX] mrp_workorder: shop floor don't override qty_producing when != 0
When marking a workorder card as done in the new shop floor display,
the qty_producing would always be overriden to the total quantity to
produce. This commit fixes this by only doing the override when the
qty_producing is zero. Also, we strike through the register production
step when the qty_producing is set to anything different from zero.

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209