[FIX] l10n_mx_edi_pos: prevent indeterministic test fail
Fix indeterministic fail of the `test_mx_pos_invoice_previous_order`
test due to an early ending of its tour. At the end of the tour, an
order is invoiced and 2 mexican fields are filled from the POS (the
`l10n_mx_edi_usage` and the `l10n_mx_edi_cfdi_to_public`), these fields
are then pushed from the client to the server and the tour ends.

Sometimes, the tour ends too early, preventing the synchronization of
the client and the server. This results in a fail when we check
`invoice.l10n_mx_edi_usage` and `invoice.l10n_mx_edi_cfdi_to_public`.

Fix: we ensure the server and the client are synchronized by ending the
tour at a 'stable' state: after invoicing the order, the button
'Invoice' is renamed to 'Reprint Invoice'.

See:
https://runbot.odoo.com/web/#id=24103&menu_id=405&cids=1&action=573&model=runbot.build.error&view_type=form

closes odoo/enterprise#45965

X-original-commit: 663f66bb4fde250f598b4e0b2cc39f383a1f28e3
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>