[FIX] website_sale_subscription: show correct subscription unit translation
Issue:
-------------------
On the website shopping when we go to a product page and there is a subscription product in the recommended product list, the unit of subscription (month, year, etc) is not being translated

Steps to reproduce:
-------------------
1. Create two products, where one of them is recurrent
2. Inside the product form -> Sales
3. Make each product be a optional/accessory product of the other
4. Publish both products
5. Go to the website shop and choose the non-recurrent product and change the language to French
6. Note that the recurrent product is displayed “par year”

Cause:
-------------------
The issue is happening because in the respective view the unit is showed by <span>per</span> and the unit is displayed by data.get('subscription_unit') which is not translated.

This makes the string “Per” be correctly translated while the unit not. To fix this issue we can use data.get('subscription_unit_display') and it correctly displays the unit using the correct translation.

OPW-3294222

closes odoo/enterprise#46962

X-original-commit: 2b409151e275db946595b32d83e43462ad3cbf22
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>