[FIX] account_sepa: keep the old way of working for the address
For the commit of the 72544175e8cfc23311aae02bd7a0b660c3b8b3ed
also for the SEPA files not corresponding to the new version of
SEPA introduced, we changed the elements we put in the address.

Sadly, some banks reject it because of this.  Before the zip and the
city were in one tag and with that commit, they were in 2.

opw-3467224

closes odoo/enterprise#45810

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>