[FIX] mrp_workorder: shop floor don't immediately update my workorders
When stopping work on a workorder card in the 'my' view of the new shop
floor display, it would immediately make the card disappear. This commit
delays this update until the data is reloaded or the currently selected
view is changed by clicking on one of the buttons in the workcenter bar.

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209