[REF] voip: complete refactoring
The code of the VoIP module is getting quite old and has accumulated a
lot of technical debt over the years, making it very difficult if not
impossible to add new features and fix some bugs. An upgrade of the code
to our current coding standards is required prior to any further
improvements.

This commit rewrites the code to use the newest APIs. In particular, it
rewrites all the javascript code in a more declarative fashion, using
Owl components and services instead of the old widgets.

Task-2376357.

closes odoo/enterprise#38339

Related: odoo/upgrade#5021
Related: odoo/odoo#115689
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>