[FIX] appointment: compute slots with default resource
Introduced in db36b59c80b45c6df1da3ac9dc876e79121e5e5e ,
Resources are now available to book via the appointment app.

This fix only concerns appointment types based on resources,
with assign_method set to 'resource_time'.

In the front-end, on the time selection page, appointment_type_page,
this assign_method translates to a dropdown above the calendar
allowing the user to select the resource. When doing so, the slots
in the calendar are updated according the availabilities of the
selected resource.

When first landing on that page, (unless a resource was selected
earlier) the first resource among the possible ones is selected.

For instance, if resources A, B, C are possible, the resource A is
selected by default and selected in the dropdown. Let us suppose
resource A is available on mondays, B on tuesdays and C on wednesdays.

BEFORE THE FIX

Despite showing resource A in the dropdown, the slots computed when
landing on the page used all possible resources.

Slots would appear on mondays, tuesdays, and wednesdays. However, as
resource A is selected in the dropdown, only mondays should contain
slots.

AFTER THE FIX

When loading the page, only use the defaut resource that will be
selected in the dropdown to compute the slots available.

Task-3460526

closes odoo/enterprise#47059

X-original-commit: d93727c2de123177af837ecdc0be9f85a750ac0a
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>