[FIX] mrp_workorder: shop floor don't mark as done twice
When marking the last workorder of an MO as done in the new shop floor
display, it shows 'close production' to also close the MO. However,
in the workorder validation we try to close the MO twice, which results
in errors that the serial number of the final product is already in use.
Hence, in this commit, the workorder validation is adapted to only call
the production validation once.

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209