[FIX] mrp_workorder: quality check ordering on WO
When opening a quality check from the new shop floor display and
validating it, the next check that was displayed was sometimes not
correct because the current_quality_check_id on the workorder was not
updated when executing quality checks on a WO out of order.
This commit makes sure the next check is always the correct one, even
when executing checks out of order.

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209