[FIX] account_asset: asset on credit note
When having an asset created from a credit note,
the acquisition date and the depreciation moves
should reflect the asset created from the reversed
move.

Steps:

- Have an account that create an asset/deferred revenue
- Create an invoice with the account set on the invoice line
- Confirm and make a credit note
-> A deferred revenue is created from the credit note, but
   the acquisition date is set to the credit note date,
   therefore the depreciation moves have a wrong date too.

With this commit, we set the acquisition date regarding the
original move, we also remove the use of `_delete_reversed_entry_assets`
to avoid deleting the asset from the invoice when it's in draft.

opw-3382294

closes odoo/enterprise#46243

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>