[REF] mrp_workorder: use unity model
Currently the shop floor app is working with 4 differents
relationalModel. It's an issue for the following points:
- It requires 4 rpc to load the server's data. Now with the
unity read, other views only require one for all differents model
- Inside the app we have to do the match between x2m objects.
We created multiple function to get the sub records base on a
lot of array browsing. With the new related key in the params,
it could be done automaticaly.
- There are reload everywhere in the code and sometimes, they
forgot to load a related object (e.g. following a public method)

This PR create a single model and clean all the reload. Using only
a single one on mrp.display (and propagate it by the env)

It also aim to have all the information on the model in a single
function. (relation, domain, order)

Minor fix on demo data, the operation are activated while the
workorders settings has been removed by default

Part-of: odoo/enterprise#46018