[FIX] hr_work_entry_contract: reload gantt view
When regenerating work entries in the gantt view, the work entries were
not reloaded, meaning that the effect of the regeneration could not be
observed unless you manually reloaded the view.

With this commit, the gantt view is automatically reloaded.

related PR that did the same for calendar views: odoo#129647

task-3382376

closes odoo/enterprise#45916

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>