[FIX] account_reports: apply search filters when printing report
Summary
-------
Currently, when a report like the General Ledger is filtered using the
search bar and then printed, the filter isn't being applied. As a
result, the printed report includes all data, not just the filtered
subset.

Steps to reproduce
------------------
* install `account_accountant`
* open the General Ledger
* filter the lines using the search bar
* print the report

You should see that everything is printed, instead of printing only the
data that matches your filter.

Cause
-----
The search filter functionality is currently only visual and doesn't
impact the backend in any way.

Fix
---
The feature used to exist, and only stopped working when reports were
converted to OWL. Thanks to this, everything necessary already exists in
the backend. We just need to update the report options to include the
`filter_search_bar` key.

opw-3424432

closes odoo/enterprise#46046

X-original-commit: ce2e591e9cfdd3ab24d5d24237ee1604ae1c7d67
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>