[FIX] mrp_workorder: shop floor move to workcenter
Since the design revamp of the new shop floor display in [1], the dialog
for workcenter selection was changed to use a checkbox view.
However, this view was also used to select a workcenter when using the
'move to workcenter' function in the WO card menu, resulting in a
traceback when trying to use this function.
This commit adapts the workcenter selection view to also support a radio
button mode next to the existing checkbox mode, and implements the
latter in the 'move to workcenter' function.

[1] https://github.com/odoo/enterprise/commit/6dd8d8e3810f6d3b22ac0dd9c61d49966fdff758

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

Part-of: odoo/enterprise#46209