[FIX] mrp_workorder: shop floor open from workorder
When pressing the open shop floor button on a workorder form view, the
shop floor view is displayed and in the search bar there is a filter set
on the MO, but it did not actually filter the records.
This commit uses search_default_name to properly filter the records.

Task: 3456817
Enterprise PR: https://github.com/odoo/enterprise/pull/45907

closes odoo/enterprise#46209

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>