[IMP] sale_subscription: allow sending the fulltraceback by mail on demand
sale_subscription can be difficult to debug when bug occurs.
This change allows to define a system parameter: 'sale_subscription.full_mail_traceback' to
send a full traceback to the responsible of the contract if something bad happens during invoice
creation. This parameter should not be activated all the times but it could make easier to
understand issues with contracts and customizations.

closes odoo/enterprise#47324

Taskid: 3459998
X-original-commit: 691b7a90847787adcd23214fa734eef9a1619faf
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>