[IMP] various: convert hr.plan to mail.activity.plan
We convert the custom plan implementation to use the generic one.

We also add a configuration menu in sale_subscription to create and edit
subscription (sale.order) plan.

See odoo/odoo#134007

Technical note:
The launch button has been removed because "launching a plan" can now be done
the same way we schedule an activity through the "Activities" button in the
chatter. That's why the test of the presence of "Launch a plan" button has been
replaced by the presence of the message of onboarding link after creating an
employee (in hr_contract_salary_tour_2).

Task-3390865

Part-of: odoo/enterprise#48599