[IMP] web_grid,timesheet_grid: hide section is no value and only one section
Before this commit, when the grid view is displayed with one section
in total and the title of that section is `None` then the section
gives the same information then the total one, which is not really
relevant.

This commit hides the section if only one section is displayed in
the grid view and the title of that section is `None`.

closes odoo/enterprise#45383

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>