[FIX] timesheet_grid: get allocated_hours of project
Before this commit, when the user goes to gantt view in My Tasks
menu and is assigned to a task linked to a project for which he
does not have access (private project) then a traceback is occured
because the progressbar in gantt view will try to fetch the
`allocated_hours` field of each project displyed in row header
of the gantt view.

Steps to reproduce the issue:
----------------------------
1. install `timesheet_grid` module (with demo data)
2. log in as Marc Demo
3. go to `Project > Tasks > My Tasks > gantt view`

Actual Behavior:
---------------
A traceback is occured saying the user cannot access to
`Research & Development` project.

Expected Behavior:
-----------------
The gantt view should load as usually without any tracebacks.

closes odoo/enterprise#45967

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>