[FIX] hr_expense_extract: fix nondeterministic failure in tests
When the `test_auto_send_for_digitization` is ran at around midnight, it
could happen that it failed because of the expense `date` not being set
to the expected value.

This was due to the fact that we rely on the `create_date` and `date` to
decide if the date should be filled or not.

Unfortunately, when `date` is set to "now" there can be
a few seconds of difference with the `create_date` leading to a mismatch
when one is set before midnight and the other after midnight.

The test will now force that `date` is set to exactly `create_date`.

closes odoo/enterprise#45993

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>