[FIX] web_studio: undeterministic invalid new app view
Before this commit, the QUnit test `web_studio > Studio > Studio Navigation: error when new app's view is invalid`
crashed undeterministically.

This was because the error flow, that is managed in studio_action_container.js triggers another owl rendering.
Sometimes the test finished before the new rendering finished, so both the error step and the contains assertions
failed. Sometimes, the test finished just before the new rendering was in the DOM, but after the error was thrown.

This commit waits for `owl.App.afterNextRender` as it is a textbook use case for it.
Local testing yields those stats:
Without Fix: crashes 4% of the time (sample: 50)
With Fix: crashes 0% of the time (sample: 150)

runbot-error-20935

closes odoo/enterprise#46060

X-original-commit: ab09ec95dce74f5c53b5589822efd4619f5b7061
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>