[FIX] timesheet_grid : fix display new line on the timesheet view
This commit's purpose is to revert a behavior that was lost on the upgrade from the saas-16.1 to the saas-16.2.
On 16.1, when a project was created on the fly on the timesheet grid page, a new active line with that project was added immediatly on the view (but it was not the case for tasks).
On 16.2, the new line was only present, even after the timer was stopped, only if the user manually reloaded the view. This commit fixes this.
In order to not trigger too many reload of the page, a reload is done
only when the timer is stopped, and the current project/task is not
corresponding to any existing lines.

task - 3471368

closes odoo/enterprise#47328

X-original-commit: b608ee5f605c09ff07809a23e98f550c6551b683
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>