[IMP] helpdesk, knowledge: adapt to "Add Properties" button changes
Purpose:
--------
In the related community PR, hiding the "Add a property" button from
the properties field and using the cog menu to add properties becomes
the normal behavior, so there is no need to create an action to add the
button in the cogmenu for each view anymore.

For the same reason, the "hideAddButton" option has been replaced in
favor of a "showAddButton" option. This commit removes the usage of the
old option and adds the new one in Knowledge (in which case there is no
cog menu).

The "fa-cog" icon in knowledge has been replaced with the "fa-cogs" icon
for consistency, as it is the icon used in the cog menu.

Task-3458627

closes odoo/enterprise#48404

Related: odoo/odoo#137589
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>