[FIX] helpdesk: fixes generate coupon performance
Before this commit, the coupon generation in helpdesk took many seconds
because a mail for the coupon generated is sent twice (when the coupon is
created and when the user send it to the customer).

This commit improves the performance of the coupon generation in helpdesk
by avoiding sending the email when the coupon is created since the user
will send it.

task-3460049

closes odoo/enterprise#50050

X-original-commit: 510b393e6197a178ce6b4d37432eb56ebc12de39
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>