[FIX] l10n_de_pos_cert: avoid checking order state when splitting order
Steps to reproduce:
1. install l10n_de_pos_res_cert
2. set up pos in German company
3. enable restaurant on pos
4. enable bill splitting
5. open pos
6. select items
7. select the split bill button
8. select item

Traceback:
```
TypeError: Cannot read properties of null (reading 'state')
```

The issue is `add_orderline` method is called in `_updateNewOrder`
where the orderline does not have an order. to avoid the issue we
should only check for the state of the order, when there is an order
in the line.

opw-3471269

closes odoo/enterprise#46373

X-original-commit: 41157050afff40479f6ae8c78a149f5d43f85e18
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>