[REM] mrp_account: remove Cost Analysis stat button in mo form
Since the 'Cost Analysis' data can be found in the MO Overview now, no
need to keep a button for it on the MO form.
Note that it's still available though actions, as it's currently still
possible to generate a Cost Analysis report based on multiple production
orders.

Task-3251506

Part-of: odoo/enterprise#40930