[FIX] account_reports: l10n_fi profit and loss sections expansion
Description of the issue/feature this commit addresses:

When using the finland localization, some lines in the profit and loss report
are shown as unfoldable while they should not either because they have no
children or because they simply don't have an expand function. This means that
when the user tries to unfold them, an error message shows up telling him
that it is not possible.

Desired behavior after the commit is merged :

This commit makes it now possible to unfold lines which do not have an
expansion function. In such a case, clicking on the line to unfold it
will simply do nothing. This way, it is still possible to have lines
without an expansion function be unfoldable.

Adding this fix, no error will be raised when unfolding a line without
an expansion function.

task-3462024

Part-of: odoo/enterprise#46417