[FIX] hr_payroll: correct distribution payment of attachments
Steps to reproduce:
-------------------
- create an employee with a contract (wage=1000)
- create two attachment for this employee:
    - A: attachment salary (monthly=200 | total=600)
    - B: attachment salary (monthly=100 | total=1000)
- create and pay a payslip for this employee

Issue:
------
Attachment A is paid with an amount of 300 instead of its monthly amount
and attachment B is not paid.

Cause:
------
When an attachment has a total amount, we try to pay its total amount
xhixh is equal to its remaining amount for the first payment
and we ignore the monthly amount.

```py
monthly_attachments = self.filtered(lambda a: not a.has_total_amount)
	fixed_total_attachments = self - monthly_attachments
....
for attachment in fixed_total_attachments:
	amount = min(attachment.remaining_amount, remaining)
```

Solution:
---------
The solution is to pay the monthly amount of all attachments first,
starting with the monthly payments (child_support).

Then, if there is still an amount to distribute,
we pay the attachments still running and which have a total amount
(starting with the earliest estimated end date).

opw-3419605

closes odoo/enterprise#46435

X-original-commit: 4fafb2983f0bb2f8eb20ca303f00610cd33c031d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>