[FIX] project_timesheet_forecast: correct user-defined filter context value
When the user analyzes the "Timesheet and Planning Analysis" pivot report from
the project module and adds it to the "Timesheet vs Planning" filter from the
Favourites at the time traceback was generated.

Steps to reproduce:
- Install the "project_timesheet_forecast" module.
- Project > Reporting > Timesheet and Planning Analysis
- Click on the pivot icon and add it to the
"Timesheet vs Planning" filter from the Favourites.
- After that, a traceback will be generated.
Error: ValueError: Invalid field 'date' on model
'project.timesheet.forecast.report.analysis'

As per the base code, the "date" invalid field existed in the user-defined
filter context , so when the user tried to analyze the
"Timesheet and Planning Analysis" pivot report and add to the
"Timesheet vs Planning" filter from the Favourites at the time, a traceback
was generated. The "date" field does not exist in the
"project.timesheet.forecast.report.analysis" object, so instead of the "date"
field in this filter, I used the "entry_date" field.

Code reference:
https://github.com/odoo/enterprise/blob/14.0/project_timesheet_forecast/report/timesheet_forecast_report.py#L16

sentry-4410474697

closes odoo/enterprise#46521

X-original-commit: c9ed630265963201d9287f6d7b01d9389f874c41
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>