[FIX] account_{accountant,reports}: populate analytic dist. on deferred entries
Steps to reproduce:
- Install Accounting
- Create a vendor bill:
  * Bill Date; 01/01/2023
  * Invoice Lines:
    - Label: [anything]
    - Account: 600000 Expenses
    - Analytic Distribution: [anything, but not empty]
    - Start Date: 01/01/2023
    - End Date: 31/12/2023
    - Price: 1200
- Confirm the bill
The lines of the created deferred entries don't have any analytic distribution.
They should have the analytic distribution of the bill that has generated them.

Solution:
1) Simply propagate the analytic distribution from the account move line
to its deferred lines if they are generated on invoice/bill validation.

2) Compute the proportion of each analytic distribution depending on
all the amounts and accounts of the deferred expenses/revenues when
the deferred lines are grouped and generated manually from
"Deferred Expense/Revenue" report.

opw-3470215

closes odoo/enterprise#47150

X-original-commit: 77f05b61cc1f145fa911966f3e9edf48edda26f1
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>