[FIX] account_base_import: fill the journal code for general journal type
Currently, when a journal is created during import, we just use the journal name as the code.
However, that gets truncated to 5 characters on create, which can conflict with existing journals.
The import should be as flexible as possible,
as it's better to let users import and fix things after the import rather than debug the import file.

If there is a journal with a shortcode that matches the truncated name of the new journal,
just use the existing journal

See also odoo/odoo#132715

closes odoo/enterprise#46101

Signed-off-by: Laurent Smet (las) <las@odoo.com>