[FIX] account_accountant: prevent resetting deferred entries to draft
Currently, when resetting an invoice to draft, we unlink/reverse its deferral moves.
This is problematic for the "Manually and Grouped" since one move should not reset the aggregated deferral of multiple moves, therefore we will not allow it anymore for this mode, only for the 'On validation' mode.

Instead, the user can create a Credit Note.

When creating the Credit Note, we currently reverse the linked deferral entries. This also a problem for the Manually & Grouped mode.
We should not alter the linked deferral entries. Instead, the created reversal will take care of creating the deferral entries (which will be reversed, therefore cancelling the original ones)

task-id 3437135

closes odoo/enterprise#46404

X-original-commit: 3419c236b8d00a0c5e90cd91c1971724b4b46e2d
Related: odoo/odoo#133330
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>