[IMP] website_studio, *: easy model exposure, website integration
*: web_studio

This commit refactors and improves the website integration proposed by Studio.
A user can now easily expose a custom model publicly on the website via the tab Website Integration.

Website pages, menus and views are created accordingly. Instead of using the notes field,
that is meant for internal use, a dedicated website description field is added on the model
when creating a page.

Two listing layouts are generated (Grid or List), as well as the single page displaying the
records individually. Those templates are used thanks to the website module that handles the
navigation and display of them.

The route on which the models are exposed is:
"/model/<string:page_name_slugified>",

With the derived routes:
"/model/<string:page_name_slugified>/page/<int:page_number>",
"/model/<string:page_name_slugified>/<string:record_slug>"

task-id-3231144

closes odoo/enterprise#47116

Related: odoo/odoo#130544
Related: odoo/upgrade#5290
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Florent Dardenne (dafl) <dafl@odoo.com>
Co-authored-by: Luca Vitali (luvi) <luvi@odoo.com>