[FIX] l10n_no_reports: Differenciating General and Local Profit and Loss
Description of the issue/feature this commit addresses:

When the Norwegian localization is installed, both the general and the
Norwegian profit and loss reports have the same name.

Desired behavior after the commit is merged :

This commit adds a country to the norwegian profit and loss report so that the
report has now a different name ("(NO)" added at the end) and it only appears
when using a Norwegian company.

task-3462024

Part-of: odoo/enterprise#46417