[FIX] l10n_syscohada_reports: Differenciating General and Local Reports
Description of the issue/feature this commit addresses:

When using the Senegalese localization and going in the Balance Sheet or the
Profit and Loss reports page, both the general and the localized reports
appear under the same name. This makes it difficult to differenciate them and
know precisely which one is open at the moment.

Desired behavior after the commit is merged :

This commit adds a "(SYSCOHADA)" after the report title so that it is now
easier to diffrenciate the two reports. The one without it is the general
one and the other is the localized one.

While the point of the task was to differenciate the general and localized
BS and P&L reports. Doing that gave the opportunity to translate them to
english and add the matching .pot and .po file for their traduction.

Adding this fix, the general Balance Sheet report and the localized one
and the general Profit and Loss report and the localized one
are easily disinguishable.

task-3462024

closes odoo/enterprise#46417

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>